drivers: sensor: hx7xx: Add new driver
Adds new driver for HX7xx based ADC chips. It supports HX710, HX711,
HX712, HX717, HX720, and HX71708. These chips are designed to measure
weight using bridge sensor, and this driver is providing an interface to
weight data and basic control (tare, calibration).
Communication with specified devices is almost identical and requires only
small changes to support all of the listed devices. This driver is
currently able to measure only the channel A. In the future will be nice
to have measurement of the channel B which can provide additional
information based on the type of ADC chip.

Devices can be distinguished by the it's package to SOP-8 and SOP-16.
SOP-8 devices do not have extra pin to control sampling frequency so it
uses communication protocol to set it. It also have fixed gain. The
SOP-16 devices have extra pins to control the sampling frequency so the
communication protocol is used to control the gain.

Signed-off-by: Tomáš Juřena <jurenatomas@gmail.com>