WIP: Enable bootloader for H7 disco board
This is just an test, not a real work.

Currently we can flash the bootloader and smp_svr project to H7 disco
board, an use mcumgr tool to manipulate the FW images. I have tested
the non-zephyr project, se we are capable of flashing any image if in
valid format and memory layout. Basicaly it means we need to place code
of the non-zephyr project to address space of slot0 and pad it by 1024
bytes of zeroes. This space will be used to header of the image later.

Signed-off-by: Tomáš Juřena <jurenatomas@gmail.com>